+++ src/phpDebug.ts
@@ -8,7 +8,7 @@ import * as childProcess from 'child_process'
 import * as path from 'path'
 import * as util from 'util'
 import * as fs from 'fs'
-import { Terminal } from './terminal'
+import { Terminal, IProgram, ProgramPidWrapper, isProcessAlive } from './terminal'
 import { convertClientPathToDebugger, convertDebuggerPathToClient, isPositiveMatchInGlobs } from './paths'
 import { minimatch } from 'minimatch'
 import { BreakpointManager, BreakpointAdapter } from './breakpoints'
@@ -120,11 +120,16 @@ export interface LaunchRequestArguments extends VSCodeDebugProtocol.LaunchReques
     env?: { [key: string]: string }
     /** Absolute path to a file containing environment variable definitions. */
     envFile?: string
-    /** If true launch the target in an external console. */
+    /** DEPRECATED: If true launch the target in an external console. */
     externalConsole?: boolean
+    /** Where to launch the debug target: internal console, integrated terminal, or external terminal. */
+    console?: 'internalConsole' | 'integratedTerminal' | 'externalTerminal'
 }
 
 class PhpDebugSession extends vscode.DebugSession {
+    /** The arguments that were given to initializeRequest */
+    private _initializeArgs: VSCodeDebugProtocol.InitializeRequestArguments
+
     /** The arguments that were given to launchRequest */
     private _args: LaunchRequestArguments
 
@@ -132,7 +137,7 @@ class PhpDebugSession extends vscode.DebugSession {
     private _server: net.Server
 
     /** The child process of the launched PHP script, if launched by the debug adapter */
-    private _phpProcess?: childProcess.ChildProcess
+    private _phpProcess?: IProgram
 
     /**
      * A map from VS Code thread IDs to Xdebug Connections.
@@ -216,6 +221,7 @@ class PhpDebugSession extends vscode.DebugSession {
         response: VSCodeDebugProtocol.InitializeResponse,
         args: VSCodeDebugProtocol.InitializeRequestArguments
     ): void {
+        this._initializeArgs = args
         response.body = {
             supportsConfigurationDoneRequest: true,
             supportsEvaluateForHovers: true,
@@ -294,18 +300,36 @@ class PhpDebugSession extends vscode.DebugSession {
             const program = args.program ? [args.program] : []
             const cwd = args.cwd || process.cwd()
             const env = Object.fromEntries(
-                Object.entries({ ...process.env, ...getConfiguredEnvironment(args) }).map(v => [
+                Object.entries(getConfiguredEnvironment(args)).map(v => [
                     v[0],
                     v[1]?.replace('${port}', port.toString()),
                 ])
             )
             // launch in CLI mode
-            if (args.externalConsole) {
-                const script = await Terminal.launchInTerminal(
-                    cwd,
-                    [runtimeExecutable, ...runtimeArgs, ...program, ...programArgs],
-                    env
-                )
+            if (args.externalConsole || args.console == 'integratedTerminal' || args.console == 'externalTerminal') {
+                let script: IProgram | undefined
+                if (this._initializeArgs.supportsRunInTerminalRequest) {
+                    const kind: 'integrated' | 'external' =
+                        args.externalConsole || args.console === 'externalTerminal' ? 'external' : 'integrated'
+                    const ritr = await new Promise<VSCodeDebugProtocol.RunInTerminalResponse>((resolve, reject) => {
+                        this.runInTerminalRequest(
+                            { args: [runtimeExecutable, ...runtimeArgs, ...program, ...programArgs], env, cwd, kind },
+                            5000,
+                            resolve
+                        )
+                    })
+                    script =
+                        ritr.success && ritr.body.shellProcessId
+                            ? new ProgramPidWrapper(ritr.body.shellProcessId)
+                            : undefined
+                } else {
+                    script = await Terminal.launchInTerminal(
+                        cwd,
+                        [runtimeExecutable, ...runtimeArgs, ...program, ...programArgs],
+                        env
+                    )
+                }
+
                 if (script) {
                     // we only do this for CLI mode. In normal listen mode, only a thread exited event is send.
                     script.on('exit', (code: number | null) => {
@@ -313,10 +337,11 @@ class PhpDebugSession extends vscode.DebugSession {
                         this.sendEvent(new vscode.TerminatedEvent())
                     })
                 }
+                // this._phpProcess = script
             } else {
                 const script = childProcess.spawn(runtimeExecutable, [...runtimeArgs, ...program, ...programArgs], {
                     cwd,
-                    env,
+                    env: { ...process.env, ...env },
                 })
                 // redirect output to debug console
                 script.stdout.on('data', (data: Buffer) => {
@@ -500,6 +525,21 @@ class PhpDebugSession extends vscode.DebugSession {
 
     private async initializeConnection(connection: xdebug.Connection): Promise<void> {
         const initPacket = await connection.waitForInitPacket()
+
+        // track the process, if we asked the IDE to spawn it
+        if (
+            !this._phpProcess &&
+            (this._args.program || this._args.runtimeArgs) &&
+            initPacket.appid &&
+            isProcessAlive(parseInt(initPacket.appid))
+        ) {
+            this._phpProcess = new ProgramPidWrapper(parseInt(initPacket.appid))
+            // we only do this for CLI mode. In normal listen mode, only a thread exited event is send.
+            this._phpProcess.on('exit', (code: number | null) => {
+                this.sendEvent(new vscode.ExitedEvent(code ?? 0))
+                this.sendEvent(new vscode.TerminatedEvent())
+            })
+        }
 
         // check if this connection should be skipped
         if (